+++ guillotina/utils/misc.py
@@ -409,12 +409,16 @@ def get_request_scheme(req) -> str:
 
 
 async def notice_on_error(key: str, func_to_await):
+    return await notice_on_error_internal(f"Error on initialize utility {key}", func_to_await)
+
+
+async def notice_on_error_internal(msg: str, func_to_await):
     try:
         await func_to_await
     except (asyncio.CancelledError, RuntimeError):
         pass
     except Exception:  # noqa
-        logger.exception(f"Error on initialize utility {key}", exc_info=True)
+        logger.exception(msg, exc_info=True)
 
 
 def dump_task_vars(pick=None) -> dict: