+++ guillotina/utils/__init__.py
@@ -35,6 +35,7 @@ from .misc import load_task_vars  # noqa
 from .misc import loop_apply_coroutine  # noqa
 from .misc import merge_dicts  # noqa
 from .misc import notice_on_error  # noqa
+from .misc import notice_on_error_internal  # noqa
 from .misc import run_async  # noqa
 from .misc import safe_unidecode  # noqa
 from .misc import strings_differ  # noqa