+++ guillotina/catalog/utils.py
@@ -12,6 +12,7 @@ from guillotina.directives import merged_tagged_value_list
 from guillotina.directives import metadata
 from guillotina.interfaces import ICatalogUtility
 from guillotina.interfaces import ISearchParser
+from guillotina.transactions import transaction
 from guillotina.utils import execute
 
 import logging
@@ -43,7 +44,12 @@ def reindex_in_future(context, security=False):
     """
     search = query_utility(ICatalogUtility)
     if search is not None:
-        execute.in_pool(search.reindex_all_content, context, security).after_request()
+        execute.in_pool(_reindex_all_content, search, context, security).after_request()
+
+
+async def _reindex_all_content(search, context, security):
+    async with transaction():
+        await search.reindex_all_content(context, security)
 
 
 _cached_indexes: typing.Dict[str, typing.Dict] = {}