+++ CHANGELOG.rst
@@ -7,6 +7,16 @@ CHANGELOG
 - Breaking change: guillotina futures won't have the txn set anymore.
   [masipcat]
 
+- Fix previous change in mailer utility
+  [masipcat]
+
+- Improve error logging in futures
+  [masipcat]
+- Documentation: Enhance search endpoint documentation
+  [ksuess]
+- Documentation: Add info about catalog_max_results
+  [ksuess]
+
 
 6.4.0rc2 (2021-11-22)
 ---------------------