+++ CHANGELOG.rst
@@ -1,8 +1,11 @@
 CHANGELOG
 =========
 
-6.4.0rc3 (unreleased)
----------------------
+6.4.0 (unreleased)
+------------------
+
+- Fix Pillow dependency
+  [bloodbare]
 
 - Breaking change: guillotina futures won't have the txn set anymore.
   [masipcat]